Removing INST_EXE
Fixing fixin so it doesn't bother trying to chmod.  The Makefile will
do that.  This also solves the problem of MM->fixin being called as a
class method causing perm_rwx to blow up under strict.

Moving INST_* constants to init_INST except for INST_MAN* which is
tied up in INSTALL_MAN*.

Fixing long standing "are we in the core" logic.

Defining PERM_RW and PERM_RWX


git-svn-id: http://svn.schwern.org/repos/CPAN/ExtUtils-MakeMaker/trunk@579 8151f2b9-fde8-0310-94fd-f048d12aab9e