From: "Craig A. Berry" <craigberry@mac.com> Date: Sat, 25 May 2002 12:19:36 -0500 To: perl5-porters@perl.org, vmsperl@perl.org, makemaker@perl.org Subject: [PATCH perl@16764/MM 5.96_01] basic.t library problem on VMS Mailing-List: contact perl5-porters-help@perl.org; run by ezmlm list-help: <mailto:perl5-porters-help@perl.org> list-unsubscribe: <mailto:perl5-porters-unsubscribe@perl.org> list-post: <mailto:perl5-porters@perl.org>
Hopefully the last VMS tweak to this twisty test.  We were using t/
as the root directory, which was preventing us from locating
libraries from ../lib (you can't go higher than the root).  This
failed to show up in earlier testing because various things in the
environment (PERL_ROOT and possibly PERL5LIB) caused the libraries to
be located, but these can't be depended on during a fresh build and
test run; they may not exist, and if they do they probably point to a
different Perl installation.  The patch below takes care of this by
making the root one level higher, which is luckily still deep enough
for the Big-Fat-Dummy tests to avoid the 8-level limit on older
systems.


git-svn-id: http://svn.schwern.org/repos/CPAN/ExtUtils-MakeMaker/trunk@793 8151f2b9-fde8-0310-94fd-f048d12aab9e