Adding BEGIN magic for core.
File::Find seems to almost always return Unix filenames, but sometimes
it blindly mixes native with Unix.  So we assume it will all be unixy and
hope that works.

Making the filename we require jive with the filename which will be put
into %INC is difficult on VMS.  If you don't get it right you get
double-loadings and lots of "subroutine redefined" warnings.

Older versions of Test::More (ie. 5.8.0) don't tolerate leaving off the
# of tests to skip.


git-svn-id: http://svn.schwern.org/repos/CPAN/ExtUtils-MakeMaker/trunk@1206 8151f2b9-fde8-0310-94fd-f048d12aab9e