* META.yml will no longer be generated in the build directory.  It will       only appear in the distdir.  This should make it easier on developers,       they don't have to worry about checking the file in all the time.     * Similarly, the SIGNATURE file will not be updated in the build       directory.  It will only be generated in the distdir.
In the course of doing this I've added the cd() method to unify handling
of "cd into this directory and run this command" in the Makefile.  Needed
it to make distdir do what I want and I wasn't going to add a bunch of
duplicate distdir methods in MM_Win95 and MM_VMS.

There's a lot of duplication yet to be eliminated using this new method.


git-svn-id: http://svn.schwern.org/repos/CPAN/ExtUtils-MakeMaker/trunk@2153 8151f2b9-fde8-0310-94fd-f048d12aab9e