- dir_target() is back.  The blibdirs target introduced in 6.18 turned       out to be not such a hot idea and is how deprecated.  Now each       directory to be created has its own target like before, but no more       .exists files.
blibdirs broke SVN's swig/perl bindings because it wanted to create
multiple different blibdirs in the same directory but the succeeding
ones thought it was all done because the blibdirs.ts file was there.

More simply, if you rm -rf blib/ MakeMaker wouldn't realize it because
the blibdirs.ts stamp was still there.

I haven't gone through and removed internal uses of blibdirs.ts yet.

Also improved the fixpath documentation to explain what the second arg
is a little better.

git-svn-id: http://svn.schwern.org/repos/CPAN/ExtUtils-MakeMaker/trunk@2220 8151f2b9-fde8-0310-94fd-f048d12aab9e