- 6.25 alphas caused a Makefile to be added to the dist.  Fixed.
We can't run the Makefile.PL in the distdir.  It leaves a Makefile around.
It could simply be deleted, but running the MPL a second time isn't a good
idea, it might be interactive.

So we have to do all the work of generating META.yml and SIGNATURE without
using a Makefile in the distdir.

git-svn-id: http://svn.schwern.org/repos/CPAN/ExtUtils-MakeMaker/trunk@2232 8151f2b9-fde8-0310-94fd-f048d12aab9e