Abuse of foreach qw() [rt.cpan.org 57124]
Zefram sez:
ExtUtils/MM_Any.pm has a line which uses foreach on an unparenthesised
qw(...) list, making use of the fake parens that the perl parser generates
for a qw(...) list.  I'm working on a core patch that would eliminate
the fake parens, deprecating this trick, and the result at the moment
is a lot of warnings due to this file.  Attached patch fixes it.