Override _fixin_replace_shebang on VMS so MM->fixin works.
Script generation doesn't look like it's ever worked on VMS, but
this hasn't been that big a deal because all the utilities
distributed with the core have their own generation facilities
under utils/.  However, Nicholas wants to put that directory on
a diet and have the scripts be generated by the modules that own
them.

So we have to get script generation via MM->fixin working, and we
do that here by prepending the same $Config{startperl} incantation
to the shebang line that we've been using in core for eons.