Generate META.json and MYMETA.json files
EU::MM now uses CPAN::Meta for all META and MYMETA generation
(both JSON v2 and YAML v1.4). This ensures output is always
valid according to the spec.

This commit removes sorting of generated metadata, as that is handled
automatically by CPAN::Meta.  The old "manual YAML" code remains for
API compatibility, but is otherwise unused.

Tests are revised to use CPAN::Meta or skip if it isn't available.

fix