Conditionally add CPAN::Meta to PREREQ_PM
CPAN::Meta is required for generating META files.  Tests will skip
without it, so not having it won't prevent installation.  However, we
don't want authors who upgrade EU::MM to lose the ability to generate
META files.  Therefore, we list CPAN::Meta in PREREQ_PM, but only if the
system EU::MM already supports INSTALL_BASE, so that dependency
resolution by a cpan client doesn't fail to install if the user is
trying to bootstrap a local library.