Redo how bundled modules are handled so they are per distribution.
This allows Scalar::Util to be special cased so it only installs if it's missing, not if
it's newer.  This avoids overwriting an XS isntall of Scalar::Util, which has functionality
the pure Perl version does not.  We just need Scalar::Util to exist.

It also avoid installing pieces of distributions, such as Scalar::Util::PP installing.

Finally, it installs more like a distribution should, all as one and based on one pm file's
version.

Closes #4