echo() and dollar sign update for VMS.
The new interface to echo() needed to be added to the override in
MM_VMS.pm.

Escaping dollar signs by doubling them doesn't do anything for the
native make utilities.  Quoting them does, so add ovverides for
escape_dollarsigns() and escape_all_dollar_signs().

In echo.t, add a space between the target name and the colon.  On
VMS, the colon is valid within a target name, so you have to have
whitespace before the colon to produce valid syntax.