Portability for my::bundles.
On VMS readdir() by default returns directory names with the .DIR
extension, which needs to be trimmed off if the directory name is
going to be glued together with other pieces to make a path.

Also on VMS, filename case is not preserved by default, so we need
to do a case-blind lookup of Scalar-List-Utils to trigger its
special handling.