Consolidate some macro handling in MM_VMS.pm.
init_others() was only handling two macros (OBJECT and LDFROM),
and both of those were also being processed (somewhat differently)
in constants().  Remove init_others() altogether and beef up what
we're doing in constants().  These macros may both consist of
space-separated lists of filenames, but in order to know that we
need to do macro expansion first, and we need to do a negative
look-behind assertion when splitting on whitespace to make sure
the whitespace is not escaped as a legal part of a modern VMS
filename.