init_dirscan case help for VMS.
Case may or may not be preserved on VMS, depending on various
versions and feature settings and so forth.  In init_dirscan, we
downcase the list of entries to be ignored, but were not doing the
same to the entries being considered, so we would fail to match
the ignore list when case is preserved.

This change makes us downcase the entries to be considered so we
are now consistent with the ignore list.