This test won't work without completely replicating pasthru().
This test can fail on ActivePerl when Strawberry Perl is installed.  ActivePerl
has a hacked %Config (see ActivePerl::Config) which will search around for an
installed make, pick up Strawberry's dmake and return that as $Config{make}
instead of nmake.

The test and pasthru() use different methods for determining what make is
in use.  To fix this we have to completely replicate pasthru() in the test
defeating the point.

A better test would do a sub module build and check that the logo is not seen.
But this isn't worth holding up the stable, so I'm just deleting this bogus test.

For #41