RT#84097 Teach MM_Any.pm to deal with VOS keep files (aka core files)
This is a bug report / suggestion against ExtUtils-MakeMaker version 6.64,
encountered while qualifying perl 5.14.2 for inclusion with OpenVOS Release 17.2.

On the Stratus VOS (aka OpenVOS) operating system, what *nix-lang calls a core file
is instead called a keep file, and it begins with the name of the program, may
include the hex form of the process ID, and ends with the suffix ".kp".  So, the
code in the clean function in MM_Any.pm needs to know of these differences.
The attached patch takes care of it. These changes have been tested by me on
OpenVOS Release 17.1 and 17.2 and the proposed changes produce the correct
behavior. Since the changes are conditioned on the 'vos' operating system tag,
they will not affect any other OS. If you have any questions or concerns,
please let me know. Thanks in advance.

Signed-off-by: Chris 'BinGOs' Williams <chris@bingosnet.co.uk>