[PATCH] Make EU::MM use LD and OPTIMIZE in recursive Makefile.PL invocations
In a directory hierarchy with Makefile.PL files in subdirectories,
command line arguments like
 perl Makefile.PL OPTIMIZE="-g -O0"
are not used in the generated subdirectory Makefiles.

This seems to be by design: there's a short list of variables
that are inherited, the rest are not taken into account at all.

Debian needs LD and OPTIMIZE for sane handling of security
related build flags, so add those to the whitelist.

Bug: http://rt.cpan.org/Public/Bug/Display.html?id=28632
Bug-Debian: http://bugs.debian.org/660195
Patch-Name: debian/makemaker-recursive.diff

Signed-off-by: Chris 'BinGOs' Williams <chris@bingosnet.co.uk>