[RT#88037] regen run of Makefile.PL will use same -MCross arg as initial .PL run
This is for the WinCE build process. If -MCross was used on the 1st run
of Makefile.PL, from buildext.pl for example, any regen runs done by
makefile, must call Makefile.PL the same way that was Makefile.PL was
initially called. Otherwise, the regen'ed makefile will be for a Desktop
Perl and not a CE Perl even though we are trying to build a CE Perl Module.

Signed-off-by: Chris 'BinGOs' Williams <chris@bingosnet.co.uk>