Remove MM_TEST_ROOT feature for VMS test support.
The feature depended on tests running in a predictable location
at a predictable depth, but since 38e7337a43b9d they don't.  Plus
the feature isn't needed on VMS systems released in the last twelve
years or so.  So it seems time to say goodbye.  Folks on VAX or VMS
versions prior to 7.2 still have years of good hunting in the
forests of EUMM 6.77 and earlier and Perl 5.18.x and earlier.  And
in fact they can still run as well as they do now but the ability
to run tests may be diminished.