VMS doesn't typically generate Unix-style Makefiles.
And in fact what gets generated by the usual non-make utilities
is a description file, not a Makefile, so let's say what gets
generated as well as what "style" it is (a dubious distinction
since MMS and MMK support the same syntax).

This is the minimum to make the new feature report something that
is not wrong on VMS and should not be construed as an endorsement
of it.  It seems questionable that we now have two lines, one of
which says "Generating..." immediately followed by another which
says "Writing...", and each of which has 50% of the information
contained in the other.