Liblist::Kid: Handle -l:foo.so invocations properly
At least on gcc and clang, -l:foo.so can be used to denote that
the actual library name is foo.so, not libfoo.so.

While nothing in the toolchain currently uses this, it'll be needed by
ExtUtils::Depend on Android; this is because Android's linker has
no concept of RTLD_GLOBAL, which means that the usual pattern of
subclassing DynaLoader and setting dl_load_flags to 1 when creating an XS
library to be used by other XS libraries won't work.

To work around that, modules on Android have to link to their
dependencies.  What this commit does is solve the issue that
XS libraries don't follow the usual naming convention of starting with
'lib', ala 'libm', so the -l switch won't find them.  Rather than teaching
MakeMaker to handle absolute paths to libraries, we can piggyback
on the code already on Liblist::Kid and simply teach it about -l: