Liblist::Kid: Libraries are not transitive on Android
On Android, if libfoo.so depends on libm.so, and you link to
libfoo, you'll only see the symbols from libm that libfoo uses;
to get access to all of libm, an explicit link is required.

This became problematic on CPAN, where for example DateTime uses
isfinite from libm, and without the -lm it would fail to spot
the symbol at runtime.

Thankfully, it appears that OS/2 has (or at least, had) the same
issue, so we can piggyback on that code.