Move MCR from PERL to PERLRUN on VMS.
And other *RUN variables.  Since on VMS you can't just supply an
executable file on the command line (well, unless it's on DCL$PATH),
but need to prepend MCR to it, we've always had PERL as a command
rather than just a filename.

But since b4de841e328911 we now have the assumption that the first
space-delimited chunk of PERL is a filename and anything after one
or more spaces are flags that can be separated out.

So move the MCR to the variables that are actually used to run
perl; I don't think think these existed when the MCR logic was
initially added, so that explains why the MCR was always in PERL.