[PATCH] Windows miniperl compatibility
On Windows, miniperl doesn’t support I/O layers, so the :utf8 binmode fails and breaks the perl build.

The attached patch should do the trick, though I have not been able to test it on Windows.

It just accomplishes the same thing but without depending on I/O layers.

Signed-off-by: Chris 'BinGOs' Williams <chris@bingosnet.co.uk>