NMake Compatibility (Part 1)
Removed the quote protectiona round the 'make' binary calls, which ultimately failed when 'nmake' was the binary being used.   The '$make' variable has additional parameters tacked onto it if 'nmake' is being used, and quoting it here causes those extra parameters to get quoted as well (rather than just the path to the binary, which I assume was the intent).