Revert "Move MCR from PERL to PERLRUN on VMS."
This reverts commit 205668e34767e990ed75e204add975b2f217e42b.

This approach isn't going to work because any random module can
specify $(PERL) in the MY namespace code expecting it to work as
a command verb rather than just a path.  For example, Encode does
that in its postamble.

So whenever we parse $(PERL) we'll just have to parse it differently
on VMS.