Handle switches in $(PERL) for VMS.
b4de841e328911 has us parsing options off $(PERL) and then adding
them back.  This has to work a bit differently on VMS because there
must be a command verb rather than just a path to the Perl
executable.

The common convention on VMS is to make "MCR" the command verb,
which basically means run this program and allow it to have
parameters without defining a foreign symbol or adding it to the
command tables.

So "MCR" is removed and then added back, but in fact we prepend it
even if it wasn't there before (for VMS only) because we always need
it and we may not have it in some cases.