Resolve the "nmake being broken" issue
There is a bug in nmake that appears if these conditions are met:
1) A command is enclosed in quotes;
2) The first parameter to that command contains a /;

nmake adds a space prior to the first / in the first parameter, if the command is in quotes.

't/*.t' was being passed as 't /*.t'

The solution is to translate '/' to '\' only when the make utility is nmake.