if ARCHLIB and LIB are identical, don't put both of them on cmd line
Previously, on Win32, the 2 are usually identical. This means every perl
process that will be launched by the make tool will have 2 -I commands to
the same directory. This makes reading what will run more difficult when
debugging a EUMM generated makefile, plus a tiny bit longer command lines
that have to be copied around by the make tool, plus the overhead of perl
parsing the extra -I on the command line. Also factor the tests out of the
loop for efficiency.