Remove fixpath call from ExtUtils::Liblist::Kid::_vms_ext.
The only thing it's accomplishing here is forcing a directory to
be a directory specification rather than a file specification,
and we have a more direct way of doing that.

Also, it's been using the long-deprecated File::Spec version of
fixpath because ExtUtils::Liblist has File::Spec in its @ISA list,
despite the comment at the top of Kid.pm stating that it must be
ISA MakeMaker.  With more testing, it might be safe to remove
File::Spec from @ISA, but that's a TODO.