Fix t/Liblist_Kid.t for VMS.
_vms_ext always returns the value of the Perl shareable library
even if the inputs are empty or non-existent, so modify the
expectations of the test for VMS only.

We use a bunch of %Config variables that are always defined in the
real world, so define them to something in the test in order to
avoid a large number of uninitialized value warnings.