Must be in t/ to call perl_lib().
The whole purpose of MakeMaker::Test::Utils::perl_lib is to set up
$ENV{PERL5LIB} relative to the t/ directory, but since 38e7337a43b
it hasn't been doing that for some tests because they've already
navigated to a temp directory before perl_lib is called.

Test::Harness, t/TEST, and t/harness mostly compensate by doing
their own mangling of $ENV{PERL5LIB}, but if you have a failing test
and run it by itself, and if it happens to run commands in child
processes (such as "perl Makefile.PL"), you'll likely be testing
the MakeMaker from your installed Perl rather than the one in the
MakeMaker distribution.

So modify those tests that chdir to a temp directory to call
perl_lib first before navigating below t/.