for dmake use makefile.mk instead of Makefile
This way nmake will never see a dmake style makefile, and wont try to run
it. dmake is less likely (but not impossible) to run a nmake style
makefile. Also it is a tiny bit faster for dmake to find the makefile,
since there won't be a failed I/O call to open "makefile.mk", then open
sucessfully "Makefile".