[RT#104663] Cygwin: do not mess with the image base and do an ephemeral rebase on i686
Using a fixed image base is a bad idea on Cygwin.

The Cygwin toolchains try to use auto-image-base wherever possible, don't mess with it.
To greatly reduce the chance of fork problems during test, do an ephemeral rebase of the just built /
installed DLL so that it doesn't conflict with other installed Cygwin libraries in i686.
The auto-image-base generally can't conflict with installed libraries on x86_64, so there's no need to do that there.
Due to a bug in binutils this must not be done when building a package with cygport or no debug information
will be extracted, so until the fix lands in Cygwin, allow the rebase to be skipped when
the environment variable CYGWIN_PACKAGE_VERSION is defined.

Signed-off-by: Chris 'BinGOs' Williams <chris@bingosnet.co.uk>