account better for bad version numbers
Older versions of EUMM just did numeric comparisons of versions with
warnings disabled.  Newer versions use version objects to do the
comparisons.  The older checks allowed some malformed versions to pass
their checks in some cases that the version objects reject.  Also, the
version listed in the files is expected to be a valid lax version, and
would crash if this wasn't the case.

Instead of rejecting these versions or crashing, do some extra massaging
to pick out a numeric prefix of a malformed version and use that.  This
should mostly match the old behavior for versions in this form, but
we're now issuing warnings about all of these cases.