MM_Unix::find_perl() dont repeatedly stat the same path in a loop
2 of the 3 branches resulted at the start of the dir loop, resulted in
the dir path never being used to compute the path of the potential perl
to stat(), and the same exact string was stat-ed in maybe_command(),
scalar(@$dirs) times. @$dirs is all the dirs in $ENV{PATH} so that is
about 1 to 3 dozen useless identical failing stats in a row. With this
change, only if the dir array will be actually used, do we run the loop
more than once. See ticket/PR for this patch for details.