Replace File::Spec->cat{dir,file} with $self->same.
This reverts f5fe6a41099c50c but there are too many conflicts to
do a simple git revert.  There was no rationale in the commit
message, so it appears to have been a cosmetic change, but it
breaks several targets on VMS.

As a grep of the sources for "sub catdir" or "sub catfile" would
have shown, the local versions are not always the same as the
File::Spec versions.

Specifically, because the parentheses in filenames are only legal
on VMS if they have been escaped, and path concatenation does this
escaping, macros must be expanded before path concatenation to
avoid getting nonsense paths such as [."$"^(INST_LIB^).auto].

That's what the local versions of catdir and catfile do on VMS and
why we have them in the first place (and have had in some form for
a couple of decades).  Other platforms will get the File::Spec
functions via inheritance as designed.