Fix XSMULTI breakage in MM_VMS::xs_make_dlsyms.
Since INST_STATIC is no longer the name of a library, but may be
a space-separated list of libraries, everything that uses it has
to be rewritten to operate on one of its components instead of the
whole thing.  This gets linker options files working.

N.B. static_lib also uses INST_STATIC and is still TODO.