Banish $(INST_STATIC) from MM_VMS.pm.
Since XSMULTI, INST_STATIC is no longer the name of a single
library but is now (at least when XSMULTI is enabled) a
space-separated list of libraries, so we needed to rewrite
everything that depended on the pre-XSMULTI design so that it
finds the individual library names some other way.

Notably, we now use alternative means to handle static libraries
one at a time both when building those libraries and when using
them to make dynamic libraries.

Note that we now use the parent static_lib and let it call our
xs_make_static_lib override.