Add overridden object files to static lib under XSMULTI/XSBUILD.
But only on VMS.  Apparently the Unix version leaves the extra
object files lying around loose and explicitly adds them to the
link command.  But there are no link changes needed if we simply
add them to the static library, which is easier to manage and
make the static library a truer reflection of the dynamic library
that will be made from it.