Get EUMM basic.t passing again on Win32 in core
There are dueling uses of PERL_CORE.  On Windows, the FIXIN macro
assumes it can run pl2bat.bat from the installed Perl if PERL_CORE
is not set, so deleting PERL_CORE from the environment as this
test was doing makes you get the wrong one (or none at all).

But if PERL_CORE is set, the metafile target is not generated in
the Makefile, so you can't test that or things that depend on it
(e.g. disttest).  So at least for now, get basic.t passing by
skipping those tests.

Signed-off-by: Chris 'BinGOs' Williams <chris@bingosnet.co.uk>