prevent EUMM::Locale from warning with old Win32.pm (#272)
* prevent EUMM::Locale from warning with old Win32.pm

Older DynaLoader inherits from AutoLoader, so nonexistent function calls
will trigger AUTOLOAD, but with a warning over deprecated behavior.

With an older Win32.pm that doesn't have the functions we are looking
for, checking for their existence with a call in an eval will trigger
this warning.  That warning breaks our tests.

Replacing the eval+call with a defined check on the function is a more
appropriate way to check for functions, and will avoid the warnings.

This unfortunately means we diverge from Encode::Locale.  Hopefully the
next version of that module will include this fix.

* Fix "unless if" syntax error