META_ADD/MERGE default meta version based on each other
Prior to commit 91b7ff8b56d529cfcee575b7858878ea0d415a0d (released in
v7.05_13), META_ADD and META_MERGE will be erroneously mixed together
when processing.  This essentially means that if a meta-spec version is
specified in one, it will be used for the other as well.

Now that META_ADD and META_MERGE are properly kept separate, they can
specify data in different spec versions.  But if a version is only
specified in one, it makes sense to use it for the other as well.  This
also matches the old behavior, even if it was at the time unintentional.