parse versions statically when possible
Parsing version declarations by evaling them is rather unsafe.  It would
be possible to improve the security of that by using a Safe container.
However, Safe.pm is an XS module and EUMM has to work without any XS.
Safe.pm should only be unavailable when compiling perl core though, and
the versions in perl core can all be parsed statically.

This patch takes the first step, by implementing static parsing of
versions that follow common patterns.  Later, a Safe container can be
added for the eval code path.  We may consider adding a warning for
the more exotic forms, and possibly rejecting them in the far future.