check specifies meta versions of META_ADD/META_MERGE before upgrading
We need the meta spec version as specified in each META_ fragment to use
as a default for the other.  Since we are upgrading the fragments
in-place, the meta version of META_ADD would always register as 2.  This
would result in META_MERGE always defaulting to being processed as 2,
unless it had a meta-spec version specified itself.

Pre-calculate the spec versions to use as defaults prior to making any
modifications to the fragments, so our upgrade process doen't affect the
expected versions.

Fixes RT#121913.