Help messup.PL mess up better.
On VMS, we can't necessarily distinguish .pl from .PL, so in
init_dirscan we peek at the contents to see if the file is a
script that generates a program.  In basic.t, the new messup.PL
was missing the hint we were looking for, so was not recognized
as a script to be run.  So add the hint.