Match final dir component in init_MANPODS
It was matching the end of the full directory spec, but on VMS,
"[foo.man1p]" does not match "man1p" at the end.  So parse out
the final directory component and match that exactly.

Also, the name of the Makefile is not necessarily 'Makefile' so
use the real name whatever it is in build_man.t