fix testing how some two versions should be parsed
Two of the versions to parse shouldn't have resulted in an eval being
done, but their expected result did not reflect that.  This didn't
result in a test failure, because the testing routine will only check if
an eval did happen when not expected, not that an eval didn't happen
when one was expected.