clean up and improve MIN_PERL_VERSION handling
Rather than having special handling for some formats in
MIN_PERL_VERSION, just use version.pm. We have a private implementation
always available even if the real version.pm isn't. This allows versions
like v5.10 or 5.008_001 to be supported.

This also adjusts the warnings and errors to be more consistent and
informative.