fix Darwin cflags override
The cflags method is what initially populates the internal CCFLAGS option
if it is not set. If we set it ourselves before calling the parent
cflags method, we don't pick up any of the values it wants to set in it.
This would prevent the hints from $Config{ccflags} being carried
through. In core, it would prevent Time::HiRes from building because of
the missing -DPERL_DARWIN option.

Follow the pattern of other OS cflags overrides and call the parent
cflags first, then modify its output and the values set.