check CPAN::Meta::Requirements capabilities rather than prereqs
CPAN::Meta::Requirements will load B to try to handle vstrings when
using pure-perl version.pm. This is an internal detail of CMR which may
not not hold true in the future. It would be preferable for it to work
without B so that CMR could be used under miniperl.

Rather than checking for B, check if CPAN::Meta::Requirements can handle
vstrings.