Test inclusion of rpath for XS modules on darwin.
Added a new test case that tests a special case for the generation of XS
modules on OS darwin. More specifically, it tests if we are able to
compile an XS module which refers to another shared library in a
non-standard location such that we can load the XS module from a perl
script without having to set the DYLD_LIBRARY_PATH environment variable.
See PR #403 and issue #402.