use a poor mans tempdir logic for parallel tests
In core we want to run the tests in parallel. The tests in EUMM
are not parallel safe as they all use the same directory for BFD
and then the teardown from one affects the other, etc. This adds
a poor mans temporary directory so each test invocation gets
its own copy of the BFD directory. This has been tested in core
builds already, and migrated to the master repo for EUMM from
there.