don't warn abut LICENSE when data comes from META file
When generating the MYMETA file, the base metadata can come from either
a META file, or if that doesn't exist the values passed to
WriteMakefile. When attempting to convert the data to a valid form, the
license may be changed to 'unknown'. If this happens, it means the
original license was invalid. But the warning issued would always refer
to the LICENSE parameter, even when the source was a META file.

Change the methods that read the META file to also return which file was
read, which can be passed along to the validation routine as the
original source of the data. If the source was a META file, skip the
warning about the LICENSE.