initialize PERL_CORE object var early and use it consistently
The PERL_CORE environment variable is used when building perl core.
Some places in the code would check for the environment variable, others
would check the PERL_CORE object attribute, and others would check a
global initialized based on the environment variable.

Move the initialization of the PERL_CORE object attribute to very early
in object construction. This allows converting all of the other users of
the setting to use the object attribute rather than taking from a
variety of sources. This makes overriding the setting work consistently
no matter how it is set.