only search for PERL_SRC when PERL_CORE is true or unset
PERL_SRC defines various things about where to build and install modules
to. If it was not specified, EUMM would search upward to find the perl
source directory. This means that for modules in the perl dist, it was
not possible to build them normally without copying the files to a
separate directory.

Change how PERL_SRC is initialized to avoid doing this search if
PERL_CORE is set to a false value. This means setting PERL_CORE to false
in your environment will allow building modules normally even if they
are in a perl source checkout. If PERL_CORE is not set, still do the
search. And if a perl source checkout is found, behave as if PERL_CORE
was true.