Take care with '.' in init_dirscan on VMS
'.' is a perfectly valid filename on VMS, denoting a file with a
zero-length name portion and a zero-length extension.  Now it isn't
a *normal* filename that one would typically use, but if something
like an errant test suite creates one in a top-level extension
directory, the Perl core build gets stuck in an endless loop on a
rerun of the build because -d '.' is true but it is not equal to
the native current directory ('[-]'), and we end up rescanning that
directory ad infinitum.

So make sure we skip it by forcing '.' to be '[-]'.