Fix pm_to_blib for VMS
154730cb0da broke the core build on VMS.  The one-liner is not
parseable and it turns out to be due to a bug in the make utility
(MMK) wherein the comment character ('#') does not get ignored
within a quoted string on a continuation line (it's fine on the
first line of a rule).  So squirrel away $#ARGV and its '#'
character in another variable on the first line, where it causes
no trouble, then use the substitute on the second line.

Also, replace literal quote characters with the q operator where
it's easy to do so; it's a lot easier to read and debug the
generated code if you don't have to think about all of the quoting
and escaping and interpolation rules regarding single and double
quotes for Perl, the make utility, and the shell all at once.