Disable XS prototypes by default
When neither -prototypes not -noprototypes is passed to xsubpp, it will
require the XS files to contain a PROTOTYPES statement. This is a
compatibility feature that dealt with a breaking change in 1996, it long
stopped having any purpose. Module::Build has disabled it since the very
first version that supported XS (0.03 in 2001), and so should we.