Can't test embedded newlines on VMS
The oneliner method is intended to build commands suitable for
writing to a Makefile or equivalent. On VMS, this works fine
because the make utility interprets the newlines with continuation
characters.  But sending that same command through backticks is
not currently working because the newlines get swallowed without
the continuation characters being consumed.  That may well be a
bug in the backticks implementation, but for now we can't reliably
use backticks to test what the make utility would do.