Use LIBS not LDFROM to link against a library
There is a newish test in 02-xsdynamic.t that tries to use LDFROM
to link against another extension.  The docs for LDFROM make it
pretty clear that it is for object files, not libraries. So switch
that test to LIBS, which also greatly simplifies the test setup
because LIBS already knows about what file extensions to use for
dynamic libraries on different platforms, etc.