Fix PERL_MM_SHEBANG when input script has args
In _fixin_replace_shebang, empty the argument string when
PERL_MM_SHEBANG=relocatable is set in the environment. Previously, if a script
had a shebang such as "#!/usr/bin/env perl" already, it would become
"#!/usr/bin/env perl perl"; on most systems, env will then try to run "perl
perl" and fail. The same issue exists if the script had a shebang with any
arguments.

Note that this occurs with the env utility before perl does its special
argument-list-as-a-string handling. No arguments can be portably preserved when
using this override.